Changed serial default, moved serial connection to server
